+++ turba/lib/Form/DeleteAddressBook.php
@@ -3,7 +3,7 @@
  * Horde_Form for deleting address books.
  *
  * See the enclosed file LICENSE for license information (ASL). If you
- * did not receive this file, see http://www.horde.org/licenses/asl.php.
+ * did not receive this file, see http://www.horde.org/licenses/apache.
  *
  * @package Turba
  */
@@ -50,14 +50,23 @@ class Turba_Form_DeleteAddressBook extends Horde_Form
             throw new Turba_Exception(_("You do not have permissions to delete this address book."));
         }
 
-        $driver = $GLOBALS['injector']->getInstance('Turba_Factory_Driver')->create($this->_addressbook->getName());
-
-        // We have a Turba_Driver, try to delete the address book.
-        $driver->deleteAll();
+        $driver = $GLOBALS['injector']
+            ->getInstance('Turba_Factory_Driver')
+            ->create($this->_addressbook->getName());
+        if ($driver->hasCapability('delete_all')) {
+            try {
+                $driver->deleteAll();
+            } catch (Turba_Exception $e) {
+                Horde::logMessage($e->getMessage(), 'ERR');
+                throw $e;
+            }
+        }
 
         // Address book successfully deleted from backend, remove the share.
         try {
-            $GLOBALS['turba_shares']->removeShare($this->_addressbook);
+            $GLOBALS['injector']
+                ->getInstance('Turba_Shares')
+                ->removeShare($this->_addressbook);
         } catch (Horde_Share_Exception $e) {
             Horde::logMessage($e->getMessage(), 'ERR');
             throw new Turba_Exception($e);
